refactor engine related stuff to its own module
