+++ rust/src/fvm/mod.rs
@@ -4,5 +4,6 @@ mod externs;
 
 pub mod machine;
 pub mod types;
+pub mod engine;
 
 pub use cgo::FvmError;
